eip_crypto: support of multiple connections of the same type
Multiple connections of the same type are supported.

For an opened file, multiple connections can be maintained. Each
connection has its state, mainly the crypto keys. The key of a
connection is cached in connection during f8/f9 request. If f8/f9
request key is the same as previous f8/f9 request of the connection,
the set key step is skipped.

EEA connection is open by QCOTA_OPEN_EEA ioctl, and EIA is open
by QCOTA_OPEN_EIA ioctl. A connection ID is returned in
qce_ota_open_conn_req structure. The connection ID is included in
qce_f8_req, or qce_f9_req to indicate which connection on behalf
of the f8 or f9 request that was issued. The connection is
released on QCOTA_CLOSE_OTA_CONN ioctl.

Change-Id: I57fb70799610a093b77ef8a1abf55f4124fda1b8
CRs-Fixed: 3637712
Signed-off-by: Che-Min Hsieh <quic_cheminh@quicinc.com>